==> notebooks/01_convert_to_hdf5.ipynb <==
import glob
import vaex
import vaex.ml
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm

hdf5_list = np.array(glob.glob('data/01_raw/hdf5/*.hdf5'))

tweets = vaex.open_many(hdf5_list)

tweets.export_hdf5

tweets.export_hdf5(path='data/02_intermediate/tweets.hdf5')
