==> notebooks/02_label_user.ipynb <==
import pandas as pd

users = pd.read_json('data/01_raw/users.jsonl', lines=True, dtype={
    "id_str": str,
})
users.set_index('id_str', inplace=True)

labels = pd.read_csv('data/01_raw/labels.csv', names=['id', 'label'], dtype={"id": str, "label": str}, index_col=0)
labels = labels.loc[~labels.index.duplicated(keep='first')]
labels.index = labels.index.astype(str)

df = pd.concat([users, labels], axis=1, join='inner')
df.index.name = 'id_str'
df.to_pickle('data/02_intermediate/users.pkl')

pd.concat([users, labels], axis=1, join='inner')
