==> notebooks/04_extract_tweet_features.ipynb <==
import vaex as vx
from kedro.io import PickleLocalDataSet
from twitter_bot_detection.helpers import log_running_time
from twitter_bot_detection.io.vaex_hdf5 import VaexHDF5DataSet

tweets = vx.open('data/02_intermediate/tweets.hdf5')

def get_dominating_hour_tweets(df):
    temp = df.groupby(['user_id', "hour"], 'count')
    temp = temp.groupby(['user_id'], agg={
        "max_hour": vx.agg.max("count")
    })
    temp = temp.to_pandas_df()
    temp.set_index("user_id", inplace=True)
    return temp

def get_dominating_dayofweek_tweets(df):
    temp = df.groupby(['user_id', "dayofweek"], 'count')
    temp = temp.groupby(['user_id'], agg={
        "max_dow": vx.agg.max("count")
    })
    temp = temp.to_pandas_df()
    temp.set_index("user_id", inplace=True)
    return temp

def get_dominating_day_of_week(df):
    temp = df.groupby(['user_id', "dayofweek"], 'count')
    temp = temp.to_pandas_df()
    temp = temp.loc[temp.groupby(['user_id'])['count'].idxmax()].drop(columns=["count"])
    return temp.set_index("user_id")

def get_dominating_hour(df):
    temp = df.groupby(['user_id', "hour"], 'count')
    temp = temp.to_pandas_df()
    temp = temp.loc[temp.groupby(['user_id'])['count'].idxmax()].drop(columns=["count"])

    return temp.set_index("user_id")
    

tweets.dtypes

@log_running_time
def extract_tweets_features(tweets: VaexHDF5DataSet) -> PickleLocalDataSet:
    features = tweets.groupby(by=tweets["user_id"], agg={
        "tweets": vx.agg.count("id"),
        "unique_sources": vx.agg.nunique("source"),
        "replies": [vx.agg.mean("is_reply")],
        "quotes": [vx.agg.mean("is_quote")],
        "hashtags": [vx.agg.mean("hashtags_count"), vx.agg.max("hashtags_count"), vx.agg.std("hashtags_count")],
        "mentions": [vx.agg.mean("mentions_count"), vx.agg.max("mentions_count"), vx.agg.std("mentions_count")],
        "urls": [vx.agg.mean("urls_count"), vx.agg.max("urls_count"), vx.agg.std("urls_count")],
        "symbols": [vx.agg.mean("symbols_count"), vx.agg.max("symbols_count"), vx.agg.std("symbols_count")],
        "sensitive": [vx.agg.mean("sensitive")],
        "truncated": [vx.agg.mean("truncated")],
        "langs": [vx.agg.nunique("lang")],
        "is_retweet": [vx.agg.mean("is_retweet")],
        #retweeted_author
        "media": [vx.agg.mean("media_count"), vx.agg.max("media_count"), vx.agg.std("media_count")],
                

#         "retweeted_by": [vx.agg.mean("retweet_count"), vx.agg.max("retweet_count"), vx.agg.std("retweet_count")],
#         "favorited_by": [vx.agg.mean("favorite_count"), vx.agg.max("favorite_count"), vx.agg.std("favorite_count")],
# #         "reply_to_tweet_ratio": vx.agg.sum("is_reply") / vx.agg.count("is_reply"),
#         "reply_to_unique": vx.agg.nunique("reply_to"),
#         "quotes_count": vx.agg.sum("is_quote"),
    })
    features = features.to_pandas_df()
    features.set_index("user_id", inplace=True)

#     features["dominating_hour"] = get_dominating_hour(tweets)
#     features["dominating_day_of_week"] = get_dominating_day_of_week(tweets)
    
#     features["dominating_hour_share"] = get_dominating_hour_tweets(tweets) / features["tweets"] 
#     features["dominating_dayofweek_share"] = get_dominating_dayofweek_tweets(tweets) / features["tweets"] 
    return features.drop(columns=["tweets"])

features = extract_tweets_features(tweets)

tweets.groupby(vx.BinnerTime.per_week(tweets['created_at']))

x = get_dominating_dayofweek_tweets(tweets)
x

x.loc[["4031355440"]]

features["dominating_hour"] = x
features

x = get_dominating_hour(tweets)
x

x = x.to_pandas_df()
x.loc[x.reset_index().groupby(['user_id'])['count'].idxmax()]

%%time
# get_dominating_hour_tweets(tweets)
x[x["user_id"] == "1000029438233448448"]

x.groupby(['user_id'], agg={
    "max_hour": vx.agg.max("count")
})

x[x["user_id"] == "2926129144"]

tweets_features = extract_tweets_features(tweets)
# tweets_features["tweets_per_lang"] = tweets_features[]

tweets_features

tweets[tweets["user_id"] == "2530039118"]["is_reply"].value_counts()

names=[
    "ID", "userID", "createdAt", "source", "isReply", "isQuote", "hashtags", 
    "hashtagsCount", "mentions", "mentionsCount", "urls", "urlsCount", 
    "symbols", "symbolsCount", "sensitive", "truncated", "lang", "isRetweet",
    "text", "retweetedAuthor", "mediaCount", "pollsCount", "replyTo", "quoteOf",
    "quoteCount", "replyCount", "retweetCount", "favoriteCount"
]

%%time
tweets = pd.read_csv('data/01_raw/tweets.csv', names=names, dtype={
    "userID": str,
    },
    usecols=[1, 2],
    parse_dates=["createdAt"],
    quoting=1,
    chunksize=1000000
)

for chunk in tweets:
    print(chunk.head(1))

%%time
len(tweets["userID"].unique())

from dask import dataframe as dd
import csv

from dask.distributed import Client, progress
client = Client()
client

%%time
dfd = dd.read_json(
    'data/01_raw/tweets.csv',
    names=names,
    dtype={
        "userID": str,
    },
#     usecols=[1, 2],
    quoting=csv.QUOTE_ALL,
    parse_dates=["createdAt"],
    lines=True
#     blocksize=128000000 # = 64 Mb chunks
).head(2000000)

%%time
dfd

%%time
import csv
x = set()
i = 0
y = None
try:
    with open('data/01_raw/test.csv', mode='rU', encoding='utf-16', newline='',) as csvfile:
        reader = csv.DictReader(csvfile, fieldnames=names, dialect='unix')
        for row in reader:
            y = row
            i += 1
            x.add(row["userID"])
except Exception as e:
    print(e)
    print(row)
len(x)
