==> notebooks/05_model_input.ipynb <==
import pandas as pd
from kedro.io import PickleLocalDataSet
from sklearn.model_selection import train_test_split
from twitter_bot_detection.helpers import log_running_time

user_features = pd.read_pickle("data/04_features/user_features.pkl")
tweets_features = pd.read_pickle("data/04_features/tweets_features.pkl")

@log_running_time
def split_dataset(user_features: PickleLocalDataSet, tweets_features: PickleLocalDataSet, test_size: str, random_state: int) -> [PickleLocalDataSet, PickleLocalDataSet, PickleLocalDataSet, PickleLocalDataSet]:
    df = pd.concat([user_features, tweets_features], axis=1, join='inner')
    X = df.drop(columns=["label"])
    y = df["label"]
    return train_test_split(X, y, test_size=test_size, random_state=random_state, stratify=y)

split_dataset(user_features, tweets_features, context.params["test_size"], context.params["random_state"])
