==> notebooks/06_automl.ipynb <==
import h2o
import pandas as pd
from h2o.automl import H2OAutoML
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, classification_report, f1_score

h2o.init()

user_features = pd.read_pickle("data/04_features/user_features.pkl")
tweets_features = pd.read_pickle("data/04_features/tweets_features.pkl")

df = pd.concat([user_features, tweets_features], axis=1, join='inner')
df.fillna(0, inplace=True)
train, test = train_test_split(df, test_size=0.2, random_state=1, stratify=df['label'])

hf = h2o.H2OFrame(train)
hf["label"] = hf["label"].asfactor()

y = 'label'
x = hf.columns

aml = H2OAutoML(seed = 1)
aml.train(x = x, y = y, training_frame = hf)

lb = aml.leaderboard
lb.head()

h2o.save_model(aml.leader, 'data/06_models/')

pred = aml.leader.predict(h2o.H2OFrame(test.drop(columns=["label"])))

print(classification_report(test["label"], pred["predict"].as_data_frame().values, digits=5))
