==> notebooks/06_xgboost.ipynb <==
import joblib
import mlflow
import numpy as np
import pandas as pd
import xgboost as xgb
from kedro.io import PickleLocalDataSet
from sklearn.preprocessing import StandardScaler
from eli5 import show_weights

from sklearn.metrics import accuracy_score, classification_report, f1_score
from sklearn.model_selection import train_test_split, StratifiedKFold

X_train = pd.read_pickle("data/05_model_input/X_train.pkl")
X_test = pd.read_pickle("data/05_model_input/X_test.pkl")
y_train = pd.read_pickle("data/05_model_input/y_train.pkl")
y_test = pd.read_pickle("data/05_model_input/y_test.pkl")

def run_xgboost(X_train: PickleLocalDataSet, X_test: PickleLocalDataSet, y_train: PickleLocalDataSet, y_test: PickleLocalDataSet, log=False) -> PickleLocalDataSet:
    params = {
        'booster': 'dart',
        'max_depth': 20, 
        'learning_rate': 0.02,
        'objective': 'binary:logistic',
        'sample_type': 'uniform',
        'normalize_type': 'tree',
        'tree_method': 'gpu_hist',
        'rate_drop': 0.1,
        'skip_drop': 0.5,
        'eval_metric': 'logloss',
        'nthread': 8,
    }
    features = X_train.columns

    xgb_train = xgb.DMatrix(X_train, label=y_train, feature_names=features)
    xgb_test = xgb.DMatrix(X_test, label=y_test, feature_names=features)
    num_round = 50
    model = xgb.train(params, xgb_train, num_round)
    y_pred = model.predict(xgb_test)

    f1 = f1_score(y_test, np.round(y_pred), average="weighted")

    joblib.dump(model, 'data/06_models/xgboost.pkl')

#     model.save_model('data/06_models/xgboost.model')
    print(classification_report(y_test, np.round(y_pred), digits=5))
    if log:
        mlflow.set_tracking_uri("databricks")
        mlflow.set_experiment(context.params.databricks)

        run_id = mlflow.search_runs(experiment_ids="3889491181315524", filter_string="tags.`mlflow.runName`='xgboost'", run_view_type=1)["run_id"][0]    
        mlflow.start_run(run_id=run_id, nested=False)
        with mlflow.start_run(nested=True):
            mlflow.set_tags({
                "lib": "xgboost",
                "description": "",
                "features": features,
            })

            mlflow.log_params(params)
            mlflow.log_metric("f1", f1, 1)
            mlflow.log_artifact('data/05_model_input/X_test.pkl')
        mlflow.end_run()
    return model

%%time
m = run_xgboost(X_train, X_test, y_train, y_test, log=True);

show_weights(m, top=50)
