==> notebooks/07_stacking.ipynb <==
import joblib
import numpy as np
from vecstack import stacking
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import ExtraTreesClassifier, RandomForestClassifier
from sklearn.metrics import accuracy_score, classification_report, f1_score

X_train = pd.read_pickle("data/05_model_input/X_train.pkl")
X_test = pd.read_pickle("data/05_model_input/X_test.pkl")
y_train = pd.read_pickle("data/05_model_input/y_train.pkl")
y_test = pd.read_pickle("data/05_model_input/y_test.pkl")

depth = 10
estimators = 100

params = {
    "iterations": 2500,
    "learning_rate": 0.02,
    "loss_function": 'Logloss',
    "random_seed": 1,
    "od_wait": 30,
    "od_type": "Iter",
    "thread_count": 8,
    "verbose": False,
    "depth": depth,
    "random_seed": 1,
}
catboost = CatBoostClassifier(**params)

xgboost = XGBClassifier(random_state=0, n_jobs=-1, learning_rate=0.1, n_estimators=estimators, max_depth=depth)
lightgbm = LGBMClassifier(random_state=0, n_jobs=-1, learning_rate=0.1, n_estimators=estimators, max_depth=depth)
extra_trees = ExtraTreesClassifier(random_state=0, n_jobs=-1,  n_estimators=estimators, max_depth=depth)
random_forest = RandomForestClassifier(random_state=0, n_jobs=-1, n_estimators=estimators, max_depth=depth)

S_train, S_test = stacking([catboost, lightgbm, xgboost, random_forest, extra_trees],
                           X_train, y_train, X_test,   # data
                           regression=False,           # classification task (if you need regression - set to True)
                           mode='oof_pred_bag',        # mode: oof for train set, predict test set in each fold and vote
                           needs_proba=False,          # predict class labels (if you need probabilities - set to True) 
                           save_dir=None,              # do not save result and log (to save in current dir - set to '.')
                           metric=f1_score,            # metric: callable
                           n_folds=4,                  # number of folds
                           stratified=True,            # stratified split for folds
                           shuffle=True,               # shuffle the data
                           random_state=0,             # ensure reproduci
                           verbose=2)                  # print all info

logreg = LogisticRegression(random_state=0).fit(S_train, y_train,)
y_pred = logreg.predict(S_test)

print(classification_report(y_test, y_pred, digits=5))
