==> notebooks/07_voting.ipynb <==
import joblib
import numpy as np
import pandas as pd
import xgboost as xgb
import seaborn as sns
from pickle import load
import matplotlib.pyplot as plt
from sklearn.linear_model import RidgeClassifierCV
from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, classification_report, f1_score

X = pd.read_pickle("data/05_model_input/X_test.pkl")
y = pd.read_pickle("data/05_model_input/y_test.pkl")

lightgbm = joblib.load('data/06_models/lightgbm.pkl')
logreg = joblib.load('data/06_models/logreg.pkl')
extra_trees = joblib.load('data/06_models/extra_trees.pkl')
catboost = joblib.load('data/06_models/catboost.pkl')
xgboost = joblib.load('data/06_models/xgboost.pkl')
random_forest = joblib.load('data/06_models/random_forest.pkl')

results = pd.DataFrame()
results["lightgbm"] = np.round(lightgbm.predict(X))
results["logreg"] = logreg.predict(X)
results["extra_trees"] = extra_trees.predict(X)
results["catboost"] = catboost.predict(X)
results["xgboost"] = np.round(xgboost.predict(xgb.DMatrix(X, label=y)))
results["random_forest"] = random_forest.predict(X)

corr = results.corr()
corr.style.background_gradient(cmap='coolwarm').set_table_attributes('style="font-size: 20px"')

print(classification_report(y, results.mode(axis=1)[0], digits=4))
